Update to coin market cap v2 api
Now that we use the v2 api, we have a loop that requests the list of
coins and stores the name and symbol in a hashmap.

When a request comes in we check the two hashmaps, and if it exists we
send a request to the api looking for the current price.