Added a save state file for remindme
The remindme plugin would lose it's data if the application crashed. Now
it is stored as a json object in a file.

As we are storing user information it is encrypted while it's being
stored in memory and doesn't fully get decrypted until it's time for the
reminder to be sent to the user. The data is also stored encrypted on
disk.

The password for the encryption is taken from an environment variable,
and a random salt is generated the first time the plugin is loaded. The
password is put through argon2 for key expansion.

Converted the Encrypted Log plugin to use argon2 instead of scrypt